Don't require OkBuffer callers to cast.
The casting masks bugs. Instead accept a long and do a range check in
OkBuffer.

+++ okcurl/src/test/java/com/squareup/okhttp/curl/MainTest.java
@@ -87,7 +87,7 @@ public class MainTest {
     try {
       OkBuffer buffer = new OkBuffer();
       body.writeTo(buffer);
-      return new String(buffer.readByteString((int) buffer.size()).toByteArray(),
+      return new String(buffer.readByteString(buffer.size()).toByteArray(),
           body.contentType().charset());
     } catch (IOException e) {
       throw new RuntimeException(e);
