+++ okio/src/test/java/okio/GzipSourceTest.java
@@ -98,8 +98,7 @@ public class GzipSourceTest {
 
   private void assertGzipped(OkBuffer gzipped) throws IOException {
     OkBuffer gunzipped = gunzip(gzipped);
-    assertEquals("It's a UNIX system! I know this!",
-        gunzipped.readUtf8((int) gunzipped.size()));
+    assertEquals("It's a UNIX system! I know this!", gunzipped.readUtf8(gunzipped.size()));
   }
 
   /**
